+++ app/scripts/settings/app-settings.types.ts
@@ -40,6 +40,7 @@ export type AppSettings = {
     has_timespan: boolean
     hits_per_page_values: number[]
     hits_per_page_default: number
+    /** Implement this to do customized async initialization. Return true to skip standard initialization afterwards. */
     initialization_checks?: (rootScope: RootScope) => Promise<boolean>
     input_case_insensitive_default?: boolean
     /** codes for translation ISO-639-1 to 639-2 */