+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Decimal separator not updated in statistics when changing language [#246](https://github.com/spraakbanken/korp-frontend/issues/246)
 - Empty value "∅" in stats table not clickable [#457](https://github.com/spraakbanken/korp-frontend/issues/457)
 - Update progress indicator when aborting search
+- Prevent bad timespan request if no corpora are available
 
 ## [9.9.0] - 2025-03-24